Add checksum for Icmp4
Since for Icmp4 packet we have to compute the checksum ourselves
the method  prep_to_send has been added to the Icmp interface.
For Icmp4 it sets the checksum value in the header and for Icmp6 it
does nothing.
I used the name prep_to_send instead of "compute_checksum" or something
alike, because compute_checksum would be missleading in the Icmp6 case.
We do not have to (and are unable to) compute the checksum for Icmp6
packet ourselves, because it has to include parts of IP header as
well. Therefore functions like send() takes care of the computation.